+++ pyflow/core/edge.py
@@ -16,7 +16,7 @@ from PyQt5.QtWidgets import (
 )
 
 from pyflow.core.serializable import Serializable
-from pyflow.graphics.socket import OCBSocket
+from pyflow.core.socket import OCBSocket
 
 
 class OCBEdge(QGraphicsPathItem, Serializable):
